organized actions into files corresponding to reducers and switched navbar when user logs in
