+++ src/components/SingleCartItem.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { Link } from 'react-router-dom'
-import { updateLineitemThunk, removeFromCartThunk } from '../redux/actions'
+import { updateLineitemThunk, removeFromCartThunk } from '../redux/actions/cart'
 import { makePriceCurrencyFormat } from '../HelperFunctions'
 
 class SingleCartItem extends Component {
@@ -117,8 +117,6 @@ const mapDispatchToProps = dispatch => {
   return {
     updateLineitem: (userId, orderId, lineitemId, lineitem) =>
       dispatch(updateLineitemThunk(userId, orderId, lineitemId, lineitem)),
-    updateProduct: (productId, product) =>
-      dispatch(updateProductThunk(productId, product)),
     removeFromCart: (userId, orderId, lineitemId) =>
       dispatch(removeFromCartThunk(userId, orderId, lineitemId))
   }
